+++ graylog2-web-interface/src/components/common/PaginatedEntityTable/slicing/slicingConstants.ts
@@ -1,0 +1,33 @@
+/*
+ * Copyright (C) 2020 Graylog, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the Server Side Public License, version 1,
+ * as published by MongoDB, Inc.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Server Side Public License for more details.
+ *
+ * You should have received a copy of the Server Side Public License
+ * along with this program. If not, see
+ * <http://www.mongodb.com/licensing/server-side-public-license>.
+ */
+
+export type SortMode = string;
+export type SortDirection = 'asc' | 'desc';
+
+export const ALPHABETICAL_SORT = 'alphabetical';
+const DEFAULT_SORT_DIRECTION: SortDirection = 'asc';
+
+export const defaultSortDirectionForMode = (
+  mode: SortMode,
+  sliceSortDefault?: { mode: SortMode; direction: SortDirection },
+): SortDirection => {
+  if (sliceSortDefault?.mode === mode) {
+    return sliceSortDefault.direction;
+  }
+
+  return DEFAULT_SORT_DIRECTION;
+};
