+++ graylog2-web-interface/src/components/events/Constants.ts
@@ -17,7 +17,7 @@
 import type { Sort, Attribute } from 'stores/PaginationTypes';
 import EventDefinitionPriorityEnum from 'logic/alerts/EventDefinitionPriorityEnum';
 import type { MiddleSectionProps } from 'components/common/PaginatedEntityTable/PaginatedEntityTable';
-import { ALPHABETICAL_SORT } from 'components/common/PaginatedEntityTable/slicing/SliceFilters';
+import { ALPHABETICAL_SORT } from 'components/common/PaginatedEntityTable/slicing/slicingConstants';
 
 export const EVENTS_ENTITY_TABLE_ID = 'events';
